requirements.txt
redis>=3.5.3
python-dotenv>=0.17.0
discord>=1.7.1